backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
pandas>=2.0,<3.0
pyarrow>=15.0,<19.0
apscheduler[sqlalchemy]>=3.10,<4.0
easyquotation>=0.7,<1.0
httpx>=0.27,<1.0
PyYAML>=6.0,<7.0
duckdb>=0.10,<2.0
requests>=2.31,<3.0
